+++ src/extension.js
@@ -30,7 +30,8 @@ const { snakeCaseTransform } = require('./commands/snake_case_transform');
 const { jsonSerializableGenerateJsonKey } = require('./commands/json_serializable_generate_json_key');
 const { createGetter } = require('./commands/create_getter');
 const { generateCopyWith } = require('./commands/generate_copy_with');
-const { generateConstructor } = require('./commands/generate_construtor');
+const { generateNamedConstructor } = require('./commands/generate_construtor');
+const { isCursorOnClassDeclaration, isCursorOnAttributeDeclaration } = require('./utils/class_content_helper');
 
 
 
@@ -88,7 +89,7 @@ function activate(context) {
 	vscode.commands.registerCommand('extension.jsonSerializableGenerateJsonKey', jsonSerializableGenerateJsonKey);
 	vscode.commands.registerCommand('extension.createGetter', createGetter);
 	vscode.commands.registerCommand('extension.generateCopyWith', generateCopyWith);
-	vscode.commands.registerCommand('extension.generateConstructor', generateConstructor);
+	vscode.commands.registerCommand('extension.generateNamedConstructor', generateNamedConstructor);
 
 	const threeTiersFolders = vscode.commands.registerCommand("extension.3-tiers", three_tiers);
 
@@ -143,18 +144,23 @@ class CodeActionProvider {
 		if (isDart3() && textFile.includes('abstract class')) {
 			codeActions.push({ command: 'extension.inheritClass', title: 'Extends Class' })
 		}
-		
+
 		const pickedText = editor.document.getText(editor.selection);
-		
-		if(pickedText.length > 2){
-			codeActions.push({ command: 'extension.snakeCaseTransform', title: 'Snake Case Transform'});
+
+		if (pickedText.length > 2) {
+			codeActions.push({ command: 'extension.snakeCaseTransform', title: 'Snake Case Transform' });
 		}
-		
-		codeActions.push({ command: 'extension.jsonSerializableGenerateJsonKey', title: 'Add JsonKey from json_serializable'});
-		codeActions.push({ command: 'extension.createGetter', title: 'Generate Getter'});
-		codeActions.push({ command: 'extension.generateCopyWith', title: 'Generate CopyWith'});
-		codeActions.push({ command: 'extension.generateConstructor', title: 'Generate Constructor with named params'});
-	
+
+		if (isCursorOnClassDeclaration()) {
+			codeActions.push({ command: 'extension.generateCopyWith', title: 'Generate CopyWith' });
+			codeActions.push({ command: 'extension.generateNamedConstructor', title: 'Generate Constructor with named params' });
+		}
+
+		if (isCursorOnAttributeDeclaration()) {
+			codeActions.push({ command: 'extension.jsonSerializableGenerateJsonKey', title: 'Add JsonKey from json_serializable' });
+			codeActions.push({ command: 'extension.createGetter', title: 'Generate Getter' });
+		}
+
 
 		if (textFile.includes(isDart3() ? 'interface' : 'abstract')) {
 			codeActions.push({
@@ -163,7 +169,7 @@ class CodeActionProvider {
 			});
 		}
 
-		
+
 
 		if (textFile === '') {
 			codeActions.push({ command: 'extension.importGist', title: 'Import GitHub Gist from id' })
@@ -220,7 +226,7 @@ class CodeActionProvider {
 					title: "Wrap with MobX Observer"
 				}
 			);
-			
+
 			codeActions.push(
 				{
 					command: "extension.fu-wrap-with-watch-signals",